% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_plotlm.R
\name{add_intervals}
\alias{add_intervals}
\title{Add a model predictions and SE band to a data.frame}
\usage{
add_intervals(data, fitted_model, SEs = 2)
}
\arguments{
\item{data}{A data.frame object with all terms used in the model}

\item{fitted_model}{a model object (ex from lm or lmer) compatible with the `predict` function}

\item{SEs}{width of confidence interval band (in se's)}
}
\value{
dataframe augmented with columns \code{y_pred}, \code{ymin}, and \code{ymax}
}
\description{
Adds columns y_pred, ymax and ymin to a data.frame based on predictions from a previously fitted
linear model (lm or lmer). Note: lmer models don't provide SEs, so those values are set to NA.
}

