% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{check_data Checks a data.frame for all the terms in a model}
\usage{
check_data(object, d = NULL)
}
\arguments{
\item{object}{Either a formula, e.g. \code{y~A+B+(1|C)}
or the output of a call to \code{lm()} or \code{lmer()}}

\item{d}{If \code{object} is a formula, provide the data.frame.
If \code{object} is the output of a model statement, leave NULL.}
}
\value{
nothing. Just prints the complete data structure.
}
\description{
Prepares a report list \code{str(data)} that displays the number of levels of each factor
including factor combinations
}
