% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable_relationships.R
\name{is_crossed}
\alias{is_crossed}
\title{Checks pairs of Variables for a crossed relationship}
\usage{
is_crossed(term1, term2, data)
}
\arguments{
\item{term1, }{One Variable name of interest.}

\item{term2}{A second Variable name of interest.}

\item{data}{data.frame with columns (minimally) \code{term1} and \code{term2}}

\item{only_fully_crossed}{= FALSE If TRUE, then will return TRUE only if the variables are fully crossed,
meaning that every level of \code{term1} is paired with every level of \code{term2} and vice versa.}
}
\value{
TRUE or FALSE
}
\description{
Given two variable names and a data.table, checks if the \code{term1} is crossed with the \code{term2}.
Two variables are at least partially crossed if at least some levels of \code{term1} are paired
with multiple levels of \code{term2} and vice versa. We consider \emph{partially crossed} a sufficient
condition for crossed relationships here. In some cases, we care about fully crossed.
}
