% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable_relationships.R
\name{is_nested}
\alias{is_nested}
\title{Checks pairs of Variables for a nested relationship}
\usage{
is_nested(outer_term, inner_term, data = NULL)
}
\arguments{
\item{outer_term}{Term expected to be at the higher level}

\item{inner_term}{Term that may be nested in \code{outer_term}}

\item{data}{data.frame with columns (minimally) \code{outer_term} and \code{inner_term}}
}
\value{
TRUE or FALSE
}
\description{
Given two variable names and a data.table, checks if the \code{inner_term} is nested in the \code{outer_term},
meaning that each level of \code{inner_term} is paired with at most 1 level of \code{outer_term}.
Each level of \code{outer_term} may be paired with multiple levels of \code{inner_term}.
}
