% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_plotlm.R
\name{stat_plotlm}
\alias{stat_plotlm}
\title{Add a trend line and SE ribbon for a model fit object}
\usage{
stat_plotlm(mapping = NULL, data = NULL, geom = "smooth",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  fitted_model = NULL, se = TRUE, fullrange = FALSE, n = 100, SEs = 2,
  ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{geom}{Use to override the default connection between
\code{geom_smooth} and \code{stat_smooth}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{fitted_model}{a model object (ex from lm or lmer) compatible with the `predict` function}

\item{se}{display confidence interval around smooth? (TRUE by default, see level to control)}

\item{fullrange}{should the fit span the full range of the plot, or just
the data}

\item{n}{number of points to evaluate smoother at}

\item{SEs}{width of confidence interval band (in se's)}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
This is a replacement for \code{stat_smooth} where you can supply your own model fit (lm or lmer)
to be plotted as a trend line.
}
\details{
The main difference from stat_smooth, or other geom_ of stat_ functions is that you must add
all of the terms in the model as Aesthetics in the aes() argument, naming them the same as in your
data: \code{aes(Group1 = Group1, Group2 = Group2, Covariate = Covariate)}
}
\examples{
library(ggplot2)
lm_mpg = lm(hwy ~ displ,mpg)
ggplot(mpg,aes(x=displ,y=hwy)) + geom_point() +
    stat_plotlm(fitted_model = lm_mpg,aes(displ = displ),se=T)

# add faceting
lm_drv = lm(hwy ~ poly(displ,2)*drv,mpg)
ggplot(mpg,aes(x=displ,y=hwy)) + geom_point() + facet_wrap(~drv,scales = 'free') +
    stat_plotlm(fitted_model = lm_drv,aes(displ = displ,drv=drv),se=T,fullrange=T,n=20)

# add multiple colors in each facet. Note that the group aesthetic is necessary for stat_plotlm
lm_trans = lm(hwy ~ poly(displ,2)*drv + trans,mpg)
ggplot(mpg,aes(x=displ,y=hwy)) + geom_point(aes(color = drv)) + facet_wrap(~trans,scales = 'free') +
   stat_plotlm(fitted_model = lm_trans,
               aes(group = drv,color = drv,displ = displ,drv=drv,trans=trans),
               se=T,fullrange=T,n=20)
}

